# Fixes issue #159 and greatly speeds up iteration of PMap in Python3