@noahbenson 2022-02-10
I should have mentioned that the commit also includes a definition for the `__reversed__` method of `PMap`. The method just throws an error saying that pmaps are not reversible. Currently if you try to reverse one, the error is somewhat arcane. (I realize that reversing a mapping like this is unusual, but you can reverse `dict` objects in Python, so making a clear error message seemed wise.)