@noahbenson 2022-02-10
Big fan of `pyrsistent` by the way—thanks for maintaining it! I use it all over my libraries [pimms](https://github.com/noahbenson/pimms) and [neuropythy](https://github.com/noahbenson/neuropythy), and I maintain a [similar library](https://github.com/noahbenson/Air.jl) in Julia!