@noahbenson 2022-02-11
Yeah, I've been using `pyrsistent` since maybe 2014? Can't quite remember but I'm still used to thinking about the data structures in a fundamentally Python2 way.

I understand removing the `itervalues()`-like methods in order to align with the direction of Python 3, but I could also see the logic in keeping them around for compatibility. (Could always stick a deprecation warning on them for awhile.) IMHO this is a somewhat aesthetic choice—I would probably eventually do this if it were my library, but I don't have a really good reason why 🤷. My main concern about these methods arose from the fact that I often pass `pyrsistent` maps into functions that expect `dict` arguments (and thus use `keys()` or `values()` and expect certain speed/behavior). I think it's fine/expected for a `dict` variant type to have extra methods/features as long as the basic dict features work as expected, though.