@tobgu 2022-02-13
Alright, lets keep the iter*-methods around for now. :-)